requirements.txt
pandas
requests
psycopg2-binary
python-dateutil
python-dotenv
